Directory stractural change:
   From: cspk/spk/lib/src
   To:   cspk/spk/spk

All tests are now refering to library header files in the cspk/spk/spk, not
the ones in /usr/local/include/spktest/ so that you don't have to
install the library while testing units.

Only one "configure.ac" is provided at the top level, cspk/spk/, so
that you can run "configure" and "make" from the top in order to
generate/build the librarary, tests and documentation with a single command.