Directory stractural change:
   From: cspk/spkpred/lib/src
   To:   cspk/spkpred/spkpred

All tests are now refering to library header files in the cspk/spkpred/spkpred,
not the ones in /usr/local/include/spktest/ so that you don't have to
install the library while testing units.

Only one "configure.ac" is provided at the top level, cspk/spkpred/, so
that you can run "configure" and "make" from the top in order to
generate/build the librarary, tests and documentation with a single command.